change the structure of the project (common and get request folder transform to Data project)
