+++ UNICS.API/Controllers/UniversityController.cs
@@ -1,12 +1,8 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using UNICS.Business.Common;
 using UNICS.Business.Services.UniversitySvc;
-using UNICS.Data.ViewModels.University;
+using UNICS.Data.ViewModels.Entities.University;
 
 namespace UNICS.API.Controllers
 {
