--- UNICS.Business/GetRequest/GetUniversityPagingRequest.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace UNICS.Business.GetRequest
-{
-    public class GetUniversityPagingRequest
-    {
-        // find by name
-        public string Name { get; set; }
-    }
-}
