+++ UNICS.Business/Services/UniversitySvc/IUniversityService.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using UNICS.Data.ViewModels.University;
+﻿using UNICS.Data.ViewModels.Entities.University;
 
 namespace UNICS.Business.Services.UniversitySvc
 {
