+++ UNICS.Data/ViewModels/Entities/Album/ViewAlbum.cs
@@ -1,12 +1,14 @@
-﻿namespace UNICS.Data.ViewModels.Major
+﻿using System;
+
+namespace UNICS.Data.ViewModels.Entities.Album
 {
-    public class ViewMajor
+    public class ViewAlbum
     {
         public int Id { get; set; }
-        public int MajorTypeId { get; set; }
         public int CompetitionId { get; set; }
+        public int AlbumTypeId { get; set; }
+        public DateTime CreateTime { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Status { get; set; }
     }
 }
