+++ UNICS.Data/ViewModels/Entities/Comment/ViewComment.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace UNICS.Data.ViewModels.Comment
+﻿namespace UNICS.Data.ViewModels.Entities.Comment
 {
     public class ViewComment
     {
