+++ UNICS.Data/ViewModels/Entities/CompetitionType/ViewCompetitionType.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace UNICS.Data.ViewModels.CompetitionType
+namespace UNICS.Data.ViewModels.Entities.CompetitionType
 {
     public class ViewCompetitionType
     {
