+++ UNICS.Data/ViewModels/Entities/Major/ViewMajor.cs
@@ -1,14 +1,12 @@
-﻿using System;
-
-namespace UNICS.Data.ViewModels.Album
+﻿namespace UNICS.Data.ViewModels.Entities.Major
 {
-    public class ViewAlbum
+    public class ViewMajor
     {
         public int Id { get; set; }
+        public int MajorTypeId { get; set; }
         public int CompetitionId { get; set; }
-        public int AlbumTypeId { get; set; }
-        public DateTime CreateTime { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
+        public int Status { get; set; }
     }
 }
