+++ UNICS.Data/ViewModels/Entities/University/GetUniversityPagingRequest.cs
@@ -1,0 +1,8 @@
+﻿namespace UNICS.Business.GetRequest
+{
+    public class GetUniversityPagingRequest
+    {
+        // find by name
+        public string Name { get; set; }
+    }
+}
